% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-mostpopular.r
\name{as.data.frame.mostpopular}
\alias{as.data.frame.mostpopular}
\title{parse mostpopular into data frame}
\usage{
\method{as.data.frame}{mostpopular}(r, ...)
}
\arguments{
\item{r}{Response object from \code{nyt_mostpopular}.}

\item{\dots}{Passed to data frame generic.}
}
\value{
Data frame with \"media\" attributes.
}
\description{
parse mostpopular into data frame
}

