% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nytapi.r
\name{as.data.frame.search}
\alias{as.data.frame.search}
\title{parse_search}
\usage{
\method{as.data.frame}{search}(x, force = TRUE, ...)
}
\arguments{
\item{force}{Logical indicating whether to force data into
a data frame. Depending on the request, a few columns may
come out ugly with recursive variables forced into non-
recursive format by separating each entry with +'s. Defaults
to TRUE.}

\item{nyt}{Response object from get_nyt}
}
\value{
Returns NYT data organized by variable.
}
\description{
Parses response data into nested list.
}
\examples{
\dontrun{
nyt <- get_nyt("political+polarization", n = 100)
nytdat <- parse_nyt(nyt)
head(nytdat)
}
}

