% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-timeswire.r
\name{as.data.frame.timeswire}
\alias{as.data.frame.timeswire}
\title{Parse nyt_timeswire object into data frame}
\usage{
\method{as.data.frame}{timeswire}(r, ...)
}
\arguments{
\item{r}{Response object returned by nyt_timeswire.}

\item{\dots}{Passed along to data.frame function.}
}
\value{
Data frame.
}
\description{
Parse nyt_timeswire object into data frame
}

