% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\name{book_list}
\alias{book_list}
\alias{ny_book_list}
\alias{ny_book_date_list}
\alias{ny_book_names}
\title{Book List}
\usage{
ny_book_list(list, bestsellers_date = NULL, published_date = NULL,
  pages = 1)

ny_book_date_list(list, published_date = NULL, pages = 1)

ny_book_names()
}
\arguments{
\item{list}{A \code{list_name_encoded} as returned by \code{ny_book_names}.}

\item{bestsellers_date}{The week-ending date for the sales reflected on \code{list-name}. 
Times best sellers lists are compiled using available book sale data. 
The \code{bestsellers_date} may be significantly earlier than published-date. 
For additional information, see the explanation at the bottom of any best-seller list page on NYTimes.com 
(example: Hardcover Fiction, published Dec. 5 but reflecting sales to Nov. 29).}

\item{published_date}{The date the best sellers list was published on NYTimes.com 
(different than bestsellers-date). Use "current" for latest list.}

\item{pages}{Number of pages of results to return.}
}
\description{
Get Best Sellers list. If no date is provided returns the latest list.
}
\examples{
\dontrun{
nytimes_key("xXXxxXxXxXXx")
books <- ny_book_names()
list <- ny_book_list(sample(books$list_name_encoded, 1))
}

}
