% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nytimes.R
\name{check_key}
\alias{check_key}
\title{Get the API key from argument, option, or prompt}
\usage{
check_key(API_KEY = NULL)
}
\arguments{
\item{API_KEY}{The API key to use. If not given, will use NYTIMES_API_KEY option or prompt}
}
\value{
the API key
}
\description{
Get the API key from argument, option, or prompt
}
