% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-mostpopular.r
\name{data.frame.mostpopular}
\alias{data.frame.mostpopular}
\title{parse mostpopular into data frame}
\usage{
data.frame.mostpopular(x, ...)
}
\arguments{
\item{x}{Response object from \code{nyt_mostpopular}.}

\item{\dots}{Passed to data frame generic.}
}
\value{
Data frame with \"media\" attributes.
}
\description{
parse mostpopular into data frame
}

