% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-timeswire.r
\name{data.frame.timeswire}
\alias{data.frame.timeswire}
\title{Parse nyt_timeswire object into data frame}
\usage{
data.frame.timeswire(...)
}
\arguments{
\item{\dots}{Passed along to data.frame function.}
}
\value{
Data frame.
}
\description{
Parse nyt_timeswire object into data frame
}

