% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nytimes.R
\name{extract_doc}
\alias{extract_doc}
\title{Convert nytimes json to tibble}
\usage{
extract_doc(doc)
}
\arguments{
\item{doc}{the nytimes json}
}
\value{
a tibble containing the relevant (for me) values
}
\description{
Convert nytimes json to tibble
}
