% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\name{history}
\alias{history}
\alias{ny_book_history}
\alias{ny_book_reviews}
\title{History & Reviews}
\usage{
ny_book_history(age_group = NULL, author = NULL, contributor = NULL,
  isbn = NULL, pages = 1, price = NULL, publisher = NULL,
  title = NULL)

ny_book_reviews(isbn = NULL, title = NULL, author = NULL)
}
\arguments{
\item{age_group}{The target age group for the best seller.}

\item{author}{The author of the best seller. 
The author field does not include additional contributors 
(see Data Structure for more details about the author and contributor fields). 
When searching the author field, you can specify any combination of first, middle and last names.}

\item{contributor}{The author of the best seller, as well as other contributors such as the illustrator 
(to search or sort by author name only, use author instead).
When searching, you can specify any combination of first, middle and last names of any of the contributors.}

\item{isbn}{nternational Standard Book Number, 10 or 13 digits. 
A best seller may have both 10-digit and 13-digit ISBNs, and may have multiple ISBNs of each type. 
To search on multiple ISBNs, separate the ISBNs with semicolons (example: 9780446579933;0061374229).}

\item{pages}{Number of pages of results to return.}

\item{price}{The publisher's list price of the best seller, including decimal point.}

\item{publisher}{The standardized name of the publisher.}

\item{title}{The title of the best seller. 
When searching, you can specify a portion of a title or a full title.}
}
\description{
Get Best Sellers list history and book reviews.
}
