% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/archive.R
\name{ny_archive}
\alias{ny_archive}
\title{Archive}
\usage{
ny_archive(year, month)
}
\arguments{
\item{year}{Year, minimum \code{1851}.}

\item{month}{Month, integer between 1 and 12.}
}
\description{
The Archive API returns an array of NYT articles for a given month, going back to 1851. 
Its response fields are the same as the Article Search API. The Archive API is very useful 
if you want to build your own database of NYT article metadata. You simply pass the API the 
year and month and it returns a JSON object with all articles for that month. 
The response size can be large (~20mb).
}
\examples{
\dontrun{
nytimes_key("xXXxxXxXxXXx")
archive <- ny_archive(2018, 1)
}

}
