% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\name{ny_book_overview}
\alias{ny_book_overview}
\title{Overview}
\usage{
ny_book_overview(published_date = NULL)
}
\arguments{
\item{published_date}{The best-seller list publication date. 
You do not have to specify the exact date the list was published. 
The service will search forward (into the future) for the closest publication date to the date you specify. 
For example, a request for \code{as.Date("2013-05-22")} will retrieve the list that was published on \code{05-26}.}
}
\description{
Get top 5 books for all the Best Sellers lists for specified date.
}
\examples{
\dontrun{
nytimes_key("xXXxxXxXxXXx")
list <- ny_book_overview(Sys.Date() - 365)
}

}
