% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geo.R
\name{ny_geo}
\alias{ny_geo}
\title{Geo}
\usage{
ny_geo(name = NULL, longitude = NULL, latitude = NULL, ne = NULL,
  sw = NULL, elevation = NULL, filter = NULL, since = NULL,
  until = NULL, facets = FALSE, sort = NULL, pages = 1)
}
\arguments{
\item{name}{A displayable name for the specified place.}

\item{longitude, latitude}{Coordinates of the specified place.}

\item{sw, ne}{Southwest and Northeast, used in combination with \code{longitude} and \code{latitude} to form a bounding box.}

\item{elevation}{The elevation of the specified place, in meters.}

\item{filter}{Filters search results based on the facets provided. For more information on the values you can filter on, see \code{facets}.}

\item{since, until}{Date range as \code{Date} objects.}

\item{facets}{When facets is set to \code{TRUE}, a count of all facets will be included in the response.}

\item{sort}{Sorts your results on the fields specified.}

\item{pages}{Number of pages of results to retrieve}
}
\description{
The Geographic API extends the Semantic API, using a linked data approach to enhance 
location concepts used in The New York Times' controlled vocabulary and data resources 
which combine them with the GeoNames database, an authoritative and free to use database 
of global geographical places, names and features.
}
