% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies.R
\name{ny_movie_reviewers}
\alias{ny_movie_reviewers}
\title{Reviewers}
\usage{
ny_movie_reviewers(reviewer = "all")
}
\arguments{
\item{reviewer}{Reviewer name or \code{all} for all reviewers, \code{full-time} for full-time reviewers, or \code{part-time} for part-time reviewers.}
}
\description{
Get movie critics.
}
\examples{
\dontrun{
nytimes_key("xXXxxXxXxXXx")
reviewers <- ny_movie_reviewers("part-time")
}

}
