% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies.R
\name{ny_movie_reviews}
\alias{ny_movie_reviews}
\title{Reviews}
\usage{
ny_movie_reviews(type = c("all", "picks"))
}
\arguments{
\item{type}{Filter by critiics, takes \code{all} or \code{picks}}
}
\description{
Get movie reviews.
}
\examples{
\dontrun{
nytimes_key("xXXxxXxXxXXx")
reviewers <- ny_movie_reviewers("part-time")
}

}
