% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/movies.R
\name{ny_movie_search}
\alias{ny_movie_search}
\title{Search Movie}
\usage{
ny_movie_search(
  q,
  picks = FALSE,
  pages = 1,
  opening_since = NULL,
  opening_until = NULL,
  order = NULL,
  publication_since = NULL,
  publication_until = NULL,
  reviewer = NULL
)
}
\arguments{
\item{q}{Search keyword (e.g. \code{lebowski}).}

\item{picks}{Set to \code{TRUE} to only show critics' picks. Otherwise shows both.}

\item{pages}{Number of pages of results to fetch, set to \code{Inf} to retrieve all.}

\item{opening_since, opening_until}{Date range as \code{Date} objects.}

\item{order}{Field to order results by (e.g. \code{by-publication-date}).}

\item{publication_since, publication_until}{Review publication date range as \code{Date} objects.}

\item{reviewer}{Filter by reviewer byline (e.g. \code{Stephen Holden}).}
}
\description{
Search for movie reviews.
}
\examples{
\dontrun{
nytimes_key("xXXxxXxXxXXx")
reviews <- ny_movie_search("wars", pages = 2)
}

}
