% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{ny_search}
\alias{ny_search}
\title{Search}
\usage{
ny_search(q, since = NULL, until = NULL, pages = 1,
  sort = c("newest", "oldest", "relevance"), facets = FALSE,
  facet_fields = NULL, facet_filter = NULL, fl = NULL, fq = NULL)
}
\arguments{
\item{q}{Search query.}

\item{since, until}{Begin and start \code{Date} objects.}

\item{pages}{Number of pages of results to return. Set to infinite (\code{Inf}) to retrieve all pages (\emph{not recommended}).}

\item{sort}{Sort order \code{newest}, \code{oldest}, \code{relevance}.}

\item{facets}{Whether to show facet counts, boolean.}

\item{facet_fields}{The following values are allowed: \code{day_of_week}, \code{document_type}, \code{ingredients}, 
\code{news_desk}, \code{pub_month}, \code{pub_year}, \code{section_name}, \code{source}, \code{subsection_name}, \code{type_of_material}.}

\item{facet_filter}{Have facet counts use filters, boolean.}

\item{fl}{List of fields to return.}

\item{fq}{Query filter.}
}
\description{
Search for NYT articles by keywords, filters and facets.
}
\examples{
\dontrun{
nytimes_key("xXXxxXxXxXXx")
trump <- ny_search("Trump", since = Sys.Date() - 3)
}

}
