% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/stories.R
\name{ny_stories}
\alias{ny_stories}
\title{Top Stories}
\usage{
ny_stories(section = "technology")
}
\arguments{
\item{section}{A valid section, one of \code{arts}, \code{automobiles}, \code{books}, 
\code{business}, \code{fashion}, \code{food}, \code{health}, \code{home}, \code{insider}, 
\code{magazine}, \code{movies}, \code{national}, \code{nyregion}, \code{obituaries}, 
\code{opinion}, \code{politics}, \code{realestate}, \code{science}, \code{sports}, 
\code{sundayreview}, \code{technology}, \code{theater}, \code{tmagazine}, \code{travel}, 
\code{upshot}, \code{world}.}
}
\description{
The Top Stories API returns an array of articles currently on the specified section.
}
\examples{
\dontrun{
nytimes_key("xXXxxXxXxXXx")
bus <- ny_stories("business")
}

}
