% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tags.R
\name{ny_tags}
\alias{ny_tags}
\title{Tags}
\usage{
ny_tags(q, filter = NULL, max = 25)
}
\arguments{
\item{q}{Your search query.}

\item{filter}{If you do not specify a value for filter (see the Optional Parameters), 
your query will be matched to tags in all four Times dictionaries: subject (\code{Des}), 
geographic location (\code{Geo}), organization (\code{Org}) and person (\code{Per}). 
You can specify moret than one in  a vector.}

\item{max}{Maximum number of results to return.}
}
\description{
With the TimesTags API, you can mine the riches of the New York Times tag set. 
The TimesTags service matches your query to the controlled vocabularies that 
fuel NYTimes.com metadata. You supply a string of characters, and the service 
returns a ranked list of suggested terms.
}
\examples{
\dontrun{
nytimes_key("xXXxxXxXxXXx")
trump <- ny_tags("Trump")
}

}
