% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-archive.r
\name{nyt_archive}
\alias{nyt_archive}
\title{Archive API}
\usage{
nyt_archive(year, month)
}
\arguments{
\item{year}{month integer integer /{year}/{month}.json}
}
\value{
Response object
}
\description{
The Archive API provides lists of NYT articles by month
  going back to 1851.
}
