% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-books.r
\name{nyt_books}
\alias{nyt_books}
\title{Books API}
\usage{
nyt_books(age_group, list, isbn, published_date, apikey, title)
}
\arguments{
\item{age_group}{author contributor isbn price publisher
title string string string string string string string
/lists/best-sellers/history.json}

\item{list}{weeks-on-list bestsellers-date date isbn
published-date rank rank-last-week offset sort-order
string integer string string string string integer
integer integer string /lists.{format}}

\item{isbn}{list-name published-date bestsellers-date
weeks-on-list rank rank-last-week offset sort-order
integer string string string integer string integer
integer string /lists/{date}/{list}.json}

\item{published_date}{api-key string string
/lists/overview.{format}}

\item{apikey}{string /lists/names.{format}}

\item{title}{isbn title author api-key integer string string
string /reviews.{format}}
}
\value{
Response object
}
\description{
The Books API provides information about book reviews and
  The New York Times bestsellers lists.
}
