% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-community.r
\name{nyt_community}
\alias{nyt_community}
\title{Community API}
\usage{
nyt_community(q = Sys.Date(), ...)
}
\arguments{
\item{q}{Type of user content query. Possible methods include
most recent comments \code{q = "recent"}, by date
\code{q = "2017-01-01"}, by user \code{q = "123498723"},
or by url
\code{q = "http://www.nytimes.com/2017/02/12/us/politics/national-security-council-turmoil.html"}.}

\item{\dots}{Passed to query in GET request.}
}
\value{
Response object
}
\description{
Get access to comments from registered users on New
  York Times articles.
}

