% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-geo.r
\name{nyt_geo}
\alias{nyt_geo}
\title{Geographic API}
\usage{
nyt_geo(name, latitude, longitude, elevation, sw, query, filter, date_range,
  facets, sort, limit, offset)
}
\arguments{
\item{name}{latitude longitude elevation sw query filter
date_range facets sort limit offset string string string
integer string string string string integer string
integer integer /query.json}
}
\value{
Response object
}
\description{
The Geographic API extends the Semantic API, using a
  linked data approach to enhance location concepts used
  in The New York Times' controlled vocabulary and data
  resources which combine them with the GeoNames database,
  an authoritative and free to use database of global
  geographical places, names and features.
}

