% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-mostpopular.r
\name{nyt_mostpopular}
\alias{nyt_mostpopular}
\title{nyt_mostpopular}
\usage{
nyt_mostpopular(metric = "mostshared", section = "all-sections", days = 7,
  apikey = NULL, ...)
}
\arguments{
\item{metric}{Popularity metric. Must be one of
\"mostemailed\" \"mostviewed\", or \"mostshared\". Defaults
to \"mostshared\".}

\item{section}{NYtimes section of interest. By default, this is
set to \"all-sections\".}

\item{days}{Window of time used to calculate metric. Must be
1, 7, or 30 (days). Defaults to 7 (or one week).}

\item{apikey}{Nytimes article search API key. By default the
function will look for the key as an environment variable.
Alternatively, you can supply the key (a character string)
directly to this apikey argument.}

\item{\dots}{Arguments passed along to query in final GET request.}
}
\value{
Nested list object of nytimes article data.
}
\description{
Retrieves NYTimes data for one of three possible
  popularity metrics---mostshared, mostviewed, or
  mostemailed.
}
