% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-search.r
\name{nyt_search}
\alias{nyt_search}
\title{nyt_search}
\usage{
nyt_search(q, n = 100, end_date = NULL, apikey = NULL, ...)
}
\arguments{
\item{q}{Search query.}

\item{n}{Number of desired articles to return.}

\item{end_date}{Date from which results should start (the results
will go back in time from this date). Must be in YYYYMMDD format,
e.g., \code{end_date = "20170101"}.}

\item{apikey}{Nytimes article search API key. By default the
function will look for the key as an environment variable.
Alternatively, you can suppply the key (a character string)
directly to this apikey argument.}

\item{\dots}{Arguments passed along to query in final GET request.}
}
\value{
Nested list object of nytimes article data.
}
\description{
Main function used to retrieve data from NYTimes'
  article search API
}

