% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt_mostpopular.r
\name{parse_mostpopular}
\alias{parse_mostpopular}
\title{parse_mostpopular}
\usage{
parse_mostpopular(r)
}
\arguments{
\item{r}{Response object from \code{nyt_mostpopular}.}
}
\value{
Data frame with \"media\" attributes.
}
\description{
parse_mostpopular
}

