% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/popular.R
\name{popular}
\alias{popular}
\alias{ny_popular_emailed}
\alias{ny_popular_shared}
\alias{ny_popular_shared_type}
\alias{ny_popular_viewed}
\title{Popular}
\usage{
ny_popular_emailed(period = 30)

ny_popular_shared(period = 30)

ny_popular_shared_type(period = 30, type = c("email", "facebook",
  "twitter"))

ny_popular_viewed(period = 30)
}
\arguments{
\item{period}{Time period: 1, 7, or 30 days.}

\item{type}{Share type: \code{email}, \code{facebook}, or \code{twitter}.}
}
\description{
Provides services for getting the most popular articles on NYTimes.com based on emails, shares, or views.
}
\section{Functions}{

\itemize{
  \item{\code{ny_popular_emailed} Returns an array of the most emailed articles on NYTimes.com for specified period of time.}
  \item{\code{ny_popular_shared} Returns an array of the most shared articles on NYTimes.com for specified period of time.}
  \item{\code{ny_popular_shared_type} Returns an array of the most shared articles by share type on NYTimes.com for specified period of time.}
  \item{\code{ny_popular_viewed} Returns an array of the most viewed articles on NYTimes.com for specified period of time.}
}
}

\examples{
\dontrun{
nytimes_key("xXXxxXxXxXXx")
emailed <- ny_popular_email(2)
}

}
