% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nytimes.R
\name{query}
\alias{query}
\title{Conduct an article search query, iterating over all pages}
\usage{
query(API_KEY = NULL, max_pages = 10, ...)
}
\arguments{
\item{API_KEY}{The API key to use. If not given, will use NYTIMES_API_KEY option or prompt}

\item{max_pages}{The maximum number of pages to retrieve (set NULL to get all pages)}

\item{...}{Options passed to API (see above)}
}
\value{
A list with meta (list) and docs (tibble) keys
}
\description{
Search the nytimes api with the given search query.
This will automatically iterate over all pages, obeying the rate limit to 10 queries / 100 articles per minute
}
\details{
To get an API key, please see https://developer.nytimes.com/.
For more information on the Article Search API, please see https://developer.nytimes.com/docs/articlesearch-product/1/overview.

Example query to get front page articles mentioning Biden or Trump on the last days of the election:
query(fq='(biden OR trump) AND source:"The New York Times" AND print_section:A AND print_page:1',
      begin_date="2020-11-01", end_date="2020-11-03")
}
