% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nytimes.R
\name{query_page}
\alias{query_page}
\title{Conduct a single query (retrieving a single page)}
\usage{
query_page(API_KEY = NULL, ...)
}
\arguments{
\item{API_KEY}{The API key to use. If not given, will use NYTIMES_API_KEY option or prompt}

\item{...}{Options passed to API (see query_pages for more details)}
}
\value{
A list with meta (list) and docs (tibble) keys
}
\description{
Conduct a single query (retrieving a single page)
}
