% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-sections.r
\name{sections}
\alias{sections}
\title{Sections of the NYtimes}
\arguments{
\item{section}{Valid entries for section arguments include
  the following:
\itemize{
    \item \"Arts\"
    \item \"Automobiles\"
    \item \"Blogs\"
    \item \"Books\"
    \item \"Business Day\"
    \item \"Education\"
    \item \"Fashion \& Style\"
    \item \"Food\"
    \item \"Health\"
    \item \"Job Market\"
    \item \"Magazine\"
    \item \"membercenter\"
    \item \"Movies\"
    \item \"Multimedia\"
    \item \"N.Y.%20%2F%20Region\"
    \item \"NYT Now\"
    \item \"Obituaries\"
    \item \"Open\"
    \item \"Opinion\"
    \item \"Public Editor\"
    \item \"Real Estate\"
    \item \"Science\"
    \item \"Sports\"
    \item \"Style\"
    \item \"Sunday Review\"
    \item \"T Magazine\"
    \item \"Technology\"
    \item \"The Upshot\"
    \item \"Theater\"
    \item \"Times Insider\"
    \item \"Today’s Paper\"
    \item \"Travel\"
    \item \"U.S.\"
    \item \"World\"
    \item \"Your Money\"
}}
}
\description{
Sections of the NYtimes
}

