% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wire.R
\name{wire}
\alias{wire}
\alias{ny_wire_content}
\alias{ny_wire_section_list}
\alias{ny_wire_source}
\alias{ny_wire_period}
\title{Wire}
\usage{
ny_wire_content(url)

ny_wire_section_list()

ny_wire_source(section, pages = 1, source = c("all", "nyt", "iht"))

ny_wire_period(section, period = 12, pages = 1, source = c("all",
  "nyt", "iht"))
}
\arguments{
\item{url}{The complete URL of a specific news item.}

\item{section}{Name of section as returned by \code{ny_wire_section_list}.}

\item{pages}{Number of pages of results to return.}

\item{source}{Name of the source; \code{all}, New York Times \code{nyt}, International New York Times \code{iht}.}

\item{period}{Limits the set of items by time published, integer in number of hours.}
}
\description{
With the Times Newswire API, you can get links and metadata for Times' 
articles as soon as they are published on NYTimes.com. The Times Newswire 
API provides an up-to-the-minute stream of published articles. You can 
filter results by source (all, nyt, or iht) and section (arts, business, ...).
}
\section{Functions}{

\itemize{
\item{\code{ny_wire_content} - Get content based on URL.} 
\item{\code{ny_wire_section_list} - Returns list of sections, useful for other calls.} 
\item{\code{ny_wire_source} - Returns list of wires by source and section.}
\item{\code{ny_wire_period} - Returns list of wires by source limited by period.} 
}
}

\examples{
\dontrun{
nytimes_key("xXXxxXxXxXXx")
sections <- ny_wire_section_list()
wires <- ny_wire_source(sample(sections$section, 1))
}

}
