% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt_mostpopular.r
\name{as.data.frame.mostpopular}
\alias{as.data.frame.mostpopular}
\title{parse_mostpopular}
\usage{
\method{as.data.frame}{mostpopular}(r, stringsAsFactors = FALSE, ...)
}
\arguments{
\item{r}{Response object from \code{nyt_mostpopular}.}
}
\value{
Data frame with \"media\" attributes.
}
\description{
parse_mostpopular
}

