% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-mostpopular.r
\name{as.data.frame.mostpopular}
\alias{as.data.frame.mostpopular}
\title{parse mostpopular into data frame}
\usage{
\method{as.data.frame}{mostpopular}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{Response object from \code{nyt_mostpopular}.}

\item{row.names}{see 'asdata.frame()'}

\item{optional}{see 'as.data.frame()' methods}

\item{\dots}{Passed to data frame generic.}
}
\value{
Data frame with \"media\" attributes.
}
\description{
parse mostpopular into data frame
}
