% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-articles.r
\name{as.data.frame.search}
\alias{as.data.frame.search}
\title{parse search}
\usage{
\method{as.data.frame}{search}(x, row.names = NULL, optional = FALSE, ...,
  force = TRUE)
}
\arguments{
\item{x}{Response object from get_nyt}

\item{row.names}{see 'asdata.frame()'}

\item{optional}{see 'as.data.frame()' methods}

\item{force}{Logical indicating whether to force data into
a data frame. Depending on the request, a few columns may
come out ugly with recursive variables forced into non-
recursive format by separating each entry with \code{+}'s.
Defaults to TRUE.}

\item{\dots}{Passed to data frame generic.}
}
\value{
Returns data frame.
}
\description{
Parses object returned by nyt_search into data frame.
}
\examples{
\dontrun{
nyt <- nyt_search("political+polarization", n = 100)
nytdat <- as.data.frame(nyt)
head(nytdat)
}
}

