% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-wire.r
\name{as.data.frame.timeswire}
\alias{as.data.frame.timeswire}
\title{Parse nyt_timeswire object into data frame}
\usage{
\method{as.data.frame}{timeswire}(x, row.names = NULL, optional = FALSE,
  ...)
}
\arguments{
\item{x}{Response object returned by nyt_timeswire.}

\item{row.names}{see 'asdata.frame()'}

\item{optional}{see 'as.data.frame()' methods}

\item{\dots}{Passed along to data.frame function.}
}
\value{
Data frame.
}
\description{
Parse nyt_timeswire object into data frame
}

