% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/books.R
\name{book_list}
\alias{book_list}
\alias{ny_book_list}
\title{Book List}
\usage{
ny_book_list(list, published_date = "current")
}
\arguments{
\item{list}{The name of the Best Sellers list.}

\item{published_date}{The date the best sellers list was published on NYTimes.com. 
Use "current" for latest list.}
}
\description{
Get Best Sellers list. If no date is provided returns the latest list.
}
\examples{
\dontrun{
nytimes_key("xXXxxXxXxXXx")
list <- ny_book_list(list = "hardcover-fiction")
}

}