% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-mostpopular.r
\name{data.frame.mostpopular}
\alias{data.frame.mostpopular}
\title{parse_mostpopular}
\usage{
data.frame.mostpopular(x, ...)
}
\arguments{
\item{r}{Response object from \code{nyt_mostpopular}.}
}
\value{
Data frame with \"media\" attributes.
}
\description{
parse_mostpopular
}

