% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/search.R
\name{ny_search}
\alias{ny_search}
\title{Search}
\usage{
ny_search(
  q,
  since = NULL,
  until = NULL,
  pages = 1,
  sort = c("newest", "oldest"),
  fq = NULL
)
}
\arguments{
\item{q}{Search query.}

\item{since, until}{Begin and start \code{Date} objects.}

\item{pages}{Number of pages of results to return. Set to \code{Inf} to retrieve all pages (up to 100).}

\item{sort}{Sort order \code{newest}, \code{oldest}.}

\item{fq}{Query filter.}
}
\description{
Search for NYT articles by keywords and filters.
}
\examples{
\dontrun{
nytimes_key("xXXxxXxXxXXx")
trump <- ny_search("Trump", since = Sys.Date() - 3)
}

}

