% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-geo.r
\name{nyt_geo}
\alias{nyt_geo}
\title{Geographic API}
\usage{
nyt_geo(q, ...)
}
\arguments{
\item{q}{Place name (vector length 1), latitude longitude,
(vector length 2), or bounding box (vector length 4). For
coordinates enter latitutde followed by longitude. For a
bounding box search, enter two pairs of lat/long coords
starting with southwestern most coordinates and ending
with northeastern most coordinates.}

\item{\dots}{Passed along to query in GET request.}
}
\value{
Response object
}
\description{
The Geographic API extends the Semantic API, using a
  linked data approach to enhance location concepts used
  in The New York Times' controlled vocabulary and data
  resources which combine them with the GeoNames database,
  an authoritative and free to use database of global
  geographical places, names and features.
}

