% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-movies.r
\name{nyt_movies}
\alias{nyt_movies}
\title{Movie Reviews API}
\usage{
nyt_movies(query = NULL, reviews = "all", critics = NULL,
  critics_pick = FALSE, ...)
}
\arguments{
\item{query}{Search string}

\item{reviews}{Name or type of reviews. Defaults to all.}

\item{critics}{Type or name of critics. Defaults to NULL.}

\item{critics_pick}{Logicl indicating whether to limit results
to top critics.}

\item{\dots}{Passed to query in get request.}
}
\value{
Response object
}
\description{
With the Movie Reviews API, you can search New York Times
  movie reviews by keyword and get lists of NYT Critics' Picks.
}
