% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nyt-timeswire.r
\name{nyt_timeswire}
\alias{nyt_timeswire}
\title{timeswire}
\usage{
nyt_timeswire(src = "all", section = "all", limit = NULL, offset = NULL,
  ...)
}
\arguments{
\item{src}{Limits the set of items by originating source
all = items from both The New York Times and The International
New York Times, nyt = New York Times items only, iht =
International New York Times items only, path. Must be one
of the following: string. Defaults to all.}

\item{section}{Limits the set of items by one or more sections
all | One or more section names, separated by semicolons.
To get all sections, specify all. To get a particular section
or sections, use section names. String. Defaults to all.}

\item{limit}{Limits the number of results, between 1 and 20,
query. Must be one of the following: integer.}

\item{offset}{Sets the starting point of the result set, query.
Must be one of the following: integer.}
}
\value{
Response object from the Times Newswire API.
}
\description{
With the Times Newswire API, you can get links and metadata
  for Times articles and blog posts as soon as they are
  published on NYTimes.com. The Times Newswire API provides
  an up-to-the-minute stream of published items.
}
\examples{
\dontrun{
nyt <- timeswire()
}
}

