% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nytapi.r
\name{parse_nyt}
\alias{parse_nyt}
\title{parse_nyt}
\usage{
parse_nyt(x)
}
\arguments{
\item{x}{Response object from get_nyt}
}
\value{
Nested list object.
}
\description{
Parses response data into nested list.
}

